Remove reflections from trace-a-ray.ray
Speed improvement here was not as large as some other gains but it was
still noticable at 50-80ms.

Old
---
"Elapsed time: 1989.896444 msecs"
"Elapsed time: 1527.073674 msecs"
"Elapsed time: 1549.851608 msecs"
"Elapsed time: 1435.910463 msecs"
"Elapsed time: 1518.823062 msecs"

New
---
"Elapsed time: 2048.542143 msecs"
"Elapsed time: 1545.355835 msecs"
"Elapsed time: 1659.81994 msecs"
"Elapsed time: 1557.797445 msecs"
"Elapsed time: 1485.270322 msecs"

One note of learning here is that when you want to do a typehint for a
record you need to require the record and you need to import it. The
import needs to use lowercase characters instead of dashes. This is
because Records are backed by Java classes and Java converts dashes to
underscores in the class path.

For example, `trace-a-ray.sphere` in the :require and
`trace_a_ray.sphere` in the :import. If I had known this I wouldn't
have put a dash in the name, perhaps. Or, maybe I would have used
underscores everywhere.

```clojure
(ns trace-a-ray.ray
  (:require [clojure.core.matrix :as m]
            [trace-a-ray.tuple :as t]
            [trace-a-ray.sphere])

  ;; Since this is a java class you need to lowercase the dashes
  ;; https://danielcompton.net/requiring-records-clojure-clojurescript
  (:import [trace_a_ray.sphere Sphere]))
```